Update index.html
The signal to noise ratio is too high at sailboat speeds to produce a useful result as coded. Working on methods to smooth, this change is a simple one that is likely too crude but what the heck, it will give us a better baseline to compare a better version to. This version includes the working Chart.js graph so we'll have a visual to help in that comparison. The 5.0
second minimum time delta filter will smooth the GPS noise some and was simple to implement as a first approach.